@@ -6,12 +6,50 @@
 
     <groupId>com.github.davidmarquis</groupId>
     <artifactId>redis-scheduler</artifactId>
-    <version>1.0.0-RELEASE</version>
+    <version>2.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
+
+    <name>redis-scheduler</name>
+    <description>Java implementation of a lightweight distributed task scheduler backed by Redis</description>
+    <url>https://github.com/davidmarquis/redis-scheduler</url>
+
+    <licenses>
+        <license>
+            <name>MIT</name>
+            <url>http://www.opensource.org/licenses/mit-license.php</url>
+            <distribution>repo</distribution>
+        </license>
+    </licenses>
+
+    <scm>
+        <connection>scm:git:git@github.com:davidmarquis/redis-scheduler.git</connection>
+        <developerConnection>scm:git:git@github.com:davidmarquis/redis-scheduler.git</developerConnection>
+        <url>git@github.com:davidmarquis/redis-scheduler.git</url>
+        <tag>HEAD</tag>
+    </scm>
+
+    <developers>
+        <developer>
+            <id>davidmarquis</id>
+            <name>David Marquis</name>
+            <email>david@radiant3.ca</email>
+        </developer>
+    </developers>
+
+    <distributionManagement>
+        <snapshotRepository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+        </snapshotRepository>
+        <repository>
+            <id>ossrh</id>
+            <url>https://oss.sonatype.org/service/local/staging/deploy/maven2/</url>
+        </repository>
+    </distributionManagement>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.6</java.version>
+        <java.version>1.8</java.version>
     </properties>
 
     <dependencies>
