@@ -1,5 +1,7 @@
 package com.github.davidmarquis.redisscheduler;
 
+import java.time.Instant;
+import java.time.LocalDateTime;
 import java.util.Calendar;
 
 /**
@@ -13,13 +15,20 @@ public interface RedisTaskScheduler {
     void setTaskTriggerListener(TaskTriggerListener taskTriggerListener);
 
     /**
+     * Runs a task immediately.
+     * @param taskId an arbitrary task identifier. That same identifier will be used in TaskTriggerListener callback
+     *               once the task is due for execution.
+     */
+    void runNow(String taskId);
+
+    /**
      * Schedules a task for future execution.
-     * @param taskId an arbitrary task identifier. That same identifier will be used in the TaskTriggerListener callback
+     * @param taskId an arbitrary task identifier. That same identifier will be used in TaskTriggerListener callback
      *               once the task is due for execution.
      * @param trigger the time at which we want the task to be executed. If this value is <code>null</code> or in the past,
      *                then the task will be immediately scheduled for execution.
      */
-    void schedule(String taskId, Calendar trigger);
+    void scheduleAt(String taskId, Instant trigger);
 
     /**
      * Removes all currently scheduled tasks from the scheduler.
